--- a/test/src/test/java/org/ethereum/beacon/test/SszTests.java
+++ b/test/src/test/java/org/ethereum/beacon/test/SszTests.java
@@ -33,6 +33,8 @@ public class SszTests extends TestUtils {
         input -> {
           SszStaticRunner testRunner = new SszStaticRunner(input.getValue0(), input.getValue1());
           return testRunner.run();
-        });
+        },
+        getRecommendedThreadCount() // run it in parallel, a lot of tests
+        );
   }
 }