--- a/test/src/test/java/org/ethereum/beacon/test/TestUtils.java
+++ b/test/src/test/java/org/ethereum/beacon/test/TestUtils.java
@@ -11,6 +11,8 @@ import org.ethereum.beacon.emulator.config.chainspec.SpecConstantsData;
 import org.ethereum.beacon.emulator.config.chainspec.SpecData;
 import org.ethereum.beacon.emulator.config.chainspec.SpecDataUtils;
 import org.ethereum.beacon.emulator.config.chainspec.SpecHelpersData;
+import org.ethereum.beacon.schedulers.Scheduler;
+import org.ethereum.beacon.schedulers.Schedulers;
 import org.ethereum.beacon.test.type.NamedTestCase;
 import org.ethereum.beacon.test.type.SpecConstantsDataMerged;
 import org.ethereum.beacon.test.type.TestCase;
@@ -30,6 +32,7 @@ import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
@@ -37,20 +40,18 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.Executors;
-import java.util.concurrent.ThreadPoolExecutor;
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.fail;
 
 public class TestUtils {
   private static final String GIT_COMMAND =
       "git submodule init & git submodule update --recursive --remote";
+  private static final Schedulers schedulers = Schedulers.createDefault();
   static ObjectMapper yamlMapper = new ObjectMapper(new YAMLFactory());
   static String PATH_TO_TESTS = "eth2.0-spec-tests/tests";
   static String PATH_TO_CONFIGS = "eth2.0-temp-test-configs";
@@ -233,15 +234,15 @@ public class TestUtils {
       Path dir,
       Class<? extends V> testsType,
       Function<Pair<TestCase, BeaconChainSpec>, Optional<String>> testCaseRunner) {
-    runTestsInResourceDirImpl(dir, testsType, testCaseRunner, Collections.emptySet(), 1);
+    runTestsInResourceDirImpl(dir, testsType, testCaseRunner, Collections.emptySet(), false);
   }
 
   static <V extends TestSkeleton> void runTestsInResourceDir(
       Path dir,
       Class<? extends V> testsType,
       Function<Pair<TestCase, BeaconChainSpec>, Optional<String>> testCaseRunner,
-      int threadNum) {
-    runTestsInResourceDirImpl(dir, testsType, testCaseRunner, Collections.emptySet(), threadNum);
+      boolean parallel) {
+    runTestsInResourceDirImpl(dir, testsType, testCaseRunner, Collections.emptySet(), parallel);
   }
 
   static <V extends TestSkeleton> void runTestsInResourceDir(
@@ -249,7 +250,7 @@ public class TestUtils {
       Class<? extends V> testsType,
       Function<Pair<TestCase, BeaconChainSpec>, Optional<String>> testCaseRunner,
       Ignored ignored) {
-    runTestsInResourceDirImpl(dir, testsType, testCaseRunner, ignored.testCases, 1);
+    runTestsInResourceDirImpl(dir, testsType, testCaseRunner, ignored.testCases, false);
   }
 
   static BeaconChainSpec loadSpecByName(String name, boolean isBlsVerified) {
@@ -282,13 +283,14 @@ public class TestUtils {
       Class<? extends V> testsType,
       Function<Pair<TestCase, BeaconChainSpec>, Optional<String>> testCaseRunner,
       Collection<String> exclusions,
-      int threadNum) {
+      boolean parallel) {
     List<File> files = getResourceFiles(dir.toString());
+    Scheduler scheduler =
+        parallel ? schedulers.cpuHeavy() : schedulers.newSingleThreadDaemon("tests");
     AtomicBoolean failed = new AtomicBoolean(false);
-    System.out.println("Running tests in " + dir + " with " + threadNum + " thread(s)");
-    ThreadPoolExecutor executor = (ThreadPoolExecutor) Executors.newFixedThreadPool(threadNum);
+    System.out.println("Running tests in " + dir + " with parallel execution set as " + parallel);
     AtomicInteger counter = new AtomicInteger(0);
-    CountDownLatch awaitLatch = new CountDownLatch(files.size());
+    List<CompletableFuture> tasks = new ArrayList<>();
     for (File file : files) {
       Runnable task =
           () -> {
@@ -301,23 +303,13 @@ public class TestUtils {
               System.out.println(num + ". \n----===----\n");
               failed.set(true);
             }
-            awaitLatch.countDown();
           };
-      executor.execute(task);
+      tasks.add(scheduler.executeR(task));
     }
 
-    try {
-      awaitLatch.await();
-    } catch (InterruptedException e) {
-      System.out.println("\n Something breaks down execution, Exception goes next.\n");
-      e.printStackTrace();
-      fail();
-    }
+    CompletableFuture[] cfs = tasks.toArray(new CompletableFuture[] {});
+    CompletableFuture.allOf(cfs).join();
     assertFalse(failed.get());
-  }
-
-  public static int getRecommendedThreadCount() {
-    return Math.max(Runtime.getRuntime().availableProcessors() / 2, 1);
   }
 
   public static class Ignored {