Following attempts were done and measured to speed up tests execution:

1. gradle default, starting point: `7m 58s`
2. gradle #threads parallel: `7m 50s`
3. gradle #cores parallel: `7m 45s`
4. gradle #cores parallel + sszTests #cores parallel: `5m 6s`
5. gradle #cores parallel + sszTests (#cores - 1) parallel : `5m 38s`
6. default gradle, sszTests #cores parallel: `5m 2s`
7. default gradle, sszTests (#threads - 1) parallel: `4m 54s`

Gradle parallel test execution is a bit tricky in skipping `wire` submodule and may be other modules in future, however it doesn't provide any sensible improvement in execution time, maybe because it runs only module by module in parallel (waiting for all tests in module are completed) or, maybe, by another reason. So, I decided to apply only sszTests parallel running with  number of threads = number of cores (6).
@mkalinin, what do you think?